+++ R/dof.R
@@ -8,7 +8,7 @@
 #' @details Methods for calculating degrees of freedom:
 #' \itemize{
 #' \item \code{"analytical"} For models of class \code{lmerMod}, Kenward-Roger approximated degrees of freedoms are calculated, for other models, \code{n-k} (number of observations minus number of parameters).
-#' \item \code{"fit"} Tries to extract residual degrees of freedom, and returns \code{Inf} if residual df's could not be extracted.
+#' \item \code{"fit"} Tries to extract residual degrees of freedom, and returns \code{Inf} if residual degrees of freedom could not be extracted.
 #' \item \code{"any"} First tries to extract residual degrees of freedom, and if these are not available, extracts analytical degrees of freedom.
 #' \item \code{"nokr"} Same as \code{"analytical"}, but does not Kenward-Roger approximation for models of class \code{lmerMod}. Instead, always uses \code{n-k} to calculate df for any model.
 #' \item \code{"wald"} Returns \code{Inf}.