+++ NEWS.md
@@ -7,6 +7,7 @@
 ## Changes to functions
 
 * `model_parameters()` now also transforms standard errors when `exponentiate = TRUE`.
+* You can now control whether `demean()` should add attributes to indicate within- and between-effects. This is only relevant for the `print()`-method of `model_parameters()`.
 
 # parameters 0.6.0